requirements.txt
aiofiles==0.6.0
alembic==1.5.4
asyncpg==0.22.0
bcrypt==3.2.0
branca==0.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
databases==0.4.1
decorator==4.4.2
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.73.0
folium==0.12.1
future==0.18.2
geocoder==1.38.1
greenlet==1.0.0
h11==0.12.0
httptools==0.1.1
idna==2.10
importlib-metadata==3.4.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.20.1
passlib==1.7.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.9.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
ratelim==0.1.6
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.17.1
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
zipp==3.4.0
